univ_agent_jac_project_enhanced/requirements.txt
jaclang
fastapi
uvicorn[standard]
requests
sentence-transformers
faiss-cpu
jaseci